Float32 downcast the SAB-backed counts + panel (halve memory, byte-identical)
Honest-note #1. The cell-major panel + gene-major counts data are raw counts (integers) -> exact in
Float32, and the kernels accumulate sums in f64, so the downcast is byte-identical (verified). Halves
the two big SABs: counts 456->228MB, panel 456->228MB (~450MB saved). toShared allows Float32Array;
dePanel/countsCSC store Float32; the worker reconstructs Float32Array; colMeanVar converts to double
internally. cores read ArrayLike<number> (unchanged).

OODA: overdispersion S100A8:206.76, DE ACTG1:-2.49917, meanVar CD3D mean 0.511507 — all identical to
the Float64 versions; SABs measured at 228MB each. 113 tests green.

+++ web/src/compute/worker.ts
@@ -7,7 +7,7 @@ import { overdispersedCore, deCore, groupStatsForCellsCore, type ODPanel } from 
 
 // Reconstruct a panel from SAB-backed buffers posted by the main thread (mapped ZERO-COPY — the buffers are shared).
 function panelFrom(p: any): ODPanel {
-  return { data: new Float64Array(p.data), indices: new Int32Array(p.indices), indptr: new Int32Array(p.indptr), nGenes: p.nGenes, lognorm: p.lognorm };
+  return { data: new Float32Array(p.data), indices: new Int32Array(p.indices), indptr: new Int32Array(p.indptr), nGenes: p.nGenes, lognorm: p.lognorm };
 }
 
 // Load the real libstar WASM kernels IN the worker (lazy, cached) — same module the main thread loads, so native C++
@@ -92,7 +92,7 @@ function runWidgetCode(args: any): Promise<any> {
     meanVar: counts ? async (lognorm = true) => {
       const M = await wasm();
       if (!M) throw new Error("WASM kernels unavailable in the worker");
-      const r = M.colMeanVar(new Float64Array(counts.data), new Int32Array(counts.indptr), counts.nCells, 1, !!lognorm);
+      const r = M.colMeanVar(new Float32Array(counts.data), new Int32Array(counts.indptr), counts.nCells, 1, !!lognorm);
       const cs = counts.symbols; const out = [];
       for (let g = 0; g < r.mean.length; g++) out.push({ symbol: cs ? cs[g] : g, mean: r.mean[g], var: r.var[g], nnz: r.nnz[g] });
       return out;
